🔧 chore: refactor imports to use absolute paths This commit refactors imports to use absolute paths instead of relative paths. This improves the readability of the code and makes it easier to maintain.
🔧 chore(server, settings, auth, main, utils, tsconfig, vite.config): refactor imports and paths
This commit refactors imports and paths to use absolute paths and aliases. This improves the readability of the code and makes it easier to maintain. The tsconfig.json file now has a baseUrl and paths property to enable the use of aliases. The vite.config.ts file now has a resolve property with an alias object to enable the use of aliases.
