+++ src/App.tsx
@@ -1,6 +1,6 @@
-import './App.css';
-import { Linka } from './components/Linka';
-import Router from './components/Router';
+import '@/App.css';
+import { Linka } from '@/components/Linka';
+import { Router } from '@/components/Router';
 
 function App() {
   const version = '2.0.0';
