+++ src/api/doShare.ts
@@ -1,5 +1,5 @@
-import { doArgs } from '../types';
-import { getConfig } from '../utils/getConfig';
+import { doArgs } from '@/types';
+import { getConfig } from '@/utils/getConfig';
 
 export async function doShare(args: doArgs) {
   const config = getConfig();
