+++ src/components/Settings/ColorModeSwitcher.tsx
@@ -1,9 +1,8 @@
-import * as React from 'react';
-
+import { ColorModeContext } from '@/contexts/ColorModeContext';
 import Brightness4Icon from '@mui/icons-material/Brightness4';
 import Brightness7Icon from '@mui/icons-material/Brightness7';
 import { Box, Button, useTheme } from '@mui/material';
-import { ColorModeContext } from '../../contexts/ColorModeContext';
+import * as React from 'react';
 
 export const ColorModeSwitcher: React.FC = () => {
   const theme = useTheme();
