+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { default as AddBookmark } from './AddBookmark';
+export { default as BookmarkForm } from './BookmarkForm';
+export { default as BookmarkFormFillButton } from './BookmarkFormFillButton';
+export { default as Credits } from './Credits';
+export { default as Linka } from './Linka';
+export { default as LinkaItem } from './LinkaItem';
+export { default as LoadingIcon } from './LoadingIcon';
+export { default as Router } from './Router';
