+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export { getConfig } from './getConfig';
+export { setConfig } from './setConfig';
