+++ src/utils/url.ts
@@ -1,7 +1,7 @@
 function shortenURL(url: string) {
-  const filterList = ["https://", "http://"];
+  const filterList = ['https://', 'http://'];
   filterList.forEach((v) => {
-    url = url.replace(v, "");
+    url = url.replace(v, '');
   });
   return url;
 }
